fix: use generic component for datetime suggestions
